Add CLI rename and enhance character prompts
